update drivers company details
