+++ src/components/DriversCompanyScheduleDetails/index.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export const DriversCompanyScheduleDetails = (props) => {
+  return (
+    <>
+      schedule
+    </>
+  )
+}
